Got it!
Alright now the s3 will hold 60fps, with the camera limited to 75,
which seems to look pretty good. It drops slightly while using the
physics, so I will have to adjust that. I think the rendering is good
now though. Will have to spend a lot of time tomorrow cleaning code.